requirements.txt
fastapi[all]>=0.103.0
httpx==0.23.0
python-dotenv==1.0.0
pydantic>=2.0.0
uvicorn
